Changes + elucidations
Added subclass ReferenceSpecimen, as a subclass of Specimen, to be used for the calilbration measurement
Added Signal class as output of the ProbeSpecimenInteraction process (currently subclass of MeasuredQuantitativeProperty, to be validated or further refined in terms of relations with EMMO)
Added PrimaryData class (subclass of EncodedData) as output of the DataPreparation, to be input for the PostProcessing
Added Secondary data as post-processed  data
Moved  Characterisation Property as subclass of Secondary data.
Elucidations, comments, examples added.