Paradigm to represent multi-techniques characterisation procedures changed
The two layered representation of CharacterisationWorkflow composed by CharacterisationMethods has been replaced by the possibility to nest CharacterisationMethods defining with sub-Methods. This allows to have more nested experiments (e.g. the automotive coatings adhesion failure industrial use case, which includes Nanoindentation and RAMAN, might become part of a new, more complex experiment), and also it is easier to represent stages that are in common for a set of subsequent methods (e.g. in a chain of techniques the sampling process is one, the specimen preparation can be one for the overall method or specific for sub methods).

CharacterisationMethod elucidation changed accordingly

CharacterisationWorkflow class removed

The object property characterisationWorkflowHasMethods has been replaced by characterisationMethodHasSubMethod