VIM definitions + Characterisation protocol and techniques taxonomy
Other definitions taken from the VIM vocabulary and specification of the VIM related term.
CharacterisationProtocol class added with an initial elucidation.
CharacterisationTechnique class added as AnalogicalIcon to logically describe a characterisation technique. It is used as the root of the characterisation techniques taxonomy (created from the NIDST characterisation method taxonomy)